Move vertex functions to separate file
No code changes, just the function definition order has been changed to
match the one from the GL.h include file.
